Canvas - resolve visibility of orphaned files (#2548)
* stashing changes

* adding more permission/hidden checks

* adding more permission/hidden checks

* test fix

* adding tests

* fix test

* missing docstrings
